+++ demo/src/app/dark/dark.component.ts
@@ -3,9 +3,13 @@ import {AppState} from "../app.state";
 
 @Component({
   selector: 'app-dark',
-  template: `<disqus [identifier]="'dark'" [shortname]="'ng2'" [removeOnDestroy]="true"></disqus>`
+  template: `<disqus [identifier]="identifier" [title]="title" [shortname]="shortname" (comment)="display($event)"></disqus>`
 })
 export class DarkComponent {
+
+  shortname = 'ng2';
+  identifier = 'dark';
+  title = 'Dark Theme';
 
   constructor(appState: AppState) {
     /** add dark class to app */
@@ -15,6 +19,10 @@ export class DarkComponent {
     appState.set('themeCover', prefixRepo("../../assets/img/dark.png"));
   }
 
+  display(comment){
+    console.log(comment);
+    alert(`It works 🤠 \n${JSON.stringify(comment)}`);
+  }
 
 }
 var prefixRepo = (path) => {