+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 #Changelog
 
+##1.1.0
+
+ - (fix) Passing identifiers, closes [#3](https://github.com/MurhafSousli/ng2-disqus/issues/3)
+ - (feat) @Output() `comment` callback (output)
+ - **[removeOnDestroy]** input is deprecated, it will remove disqus script on destroy by default.
+ 
+
 ##1.0.4
  - Improve component code
  - Add URL validator